Add debug mode, smoother start, title bar and design revamp
This commit includes the new debug mode, use "npm run-script debug" to run it
or "--debug-mode" with the packaged app.

With debug mode implemented, I had implemented smooth start, which shows the window after
the HTML finished rendering.

A new titlebar was included since on macOS, the status bar would behave weirdly with overlays.
It's implemented on macOS, but not tested under Windows and Linux yet.

A design revamp is underway, I want the player to feel more *dynamic* with status and songs.
Gradients probably don't help. I'll what to do with this.

Unstaged, is some additional code not ready yet.