I've changed my mind.
The project will resume as it were before clearing the repository.
It will likely have to be refactored after MVC.

I will certainly need some help when it comes to coming
up with a better way to lay out the HTML, CSS and JS files.
Possibly migrating to a React.js, or Anglular and some
transpilers (hopefully LESS and TypeScript) to better keep
things.