Rewrite the player system
The player system was rewritten to allow for player "cores", modules which
allow for songs to come from multiple providers like Spotify (if I can get
Spotify Connect to work) as opposed to just using the HTML5 Audio API.

The tasks for maintaining a playlist, as well as managing which song is
playing when stays within itself (player.js) the actual playing is done
by the core (player-html.js).

I plan to just ship with HTML5 Audio for now and add more later on. I'll also
be looking at seeing if there is a libav bind for Node.js, allowing more file
formats to be played.